Added file watcher. Attempted to separated UI and logic (and failed).

+++ ParallelReality/ModManager.cs
@@ -1,21 +1,41 @@
 /*  ModManager.cs
- *  Version 1.4 (2025.04.13)
+ *  Version 1.5 (2025.04.16)
  *  
  *  Contributor
  *      Arime-chan (Author)
  */
+
+#define DEV
+
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
+using ExShared;
 
 namespace ParallelReality
 {
-    internal class ModManager
+    public enum ModManagerState : byte
     {
-        public const string STR_DATA_FOLDER_NAME = "Reality Break_Data";
+        Idle,
+        CheckingForMods,
+        BackingUpBaseGame,
+        RestoringBaseGame,
+        ApplyingMods,
+    }
+
+    public class ModManager
+    {
+        public const string STR_FOLDER_NAME_DATA = "Reality Break_Data";
+        public const string STR_FOLDER_NAME_STREAMING_ASSETS = "StreamingAssets";
+        public const string STR_PATH_DATA_STREAMING_ASSETS = STR_FOLDER_NAME_DATA + "/" + STR_FOLDER_NAME_STREAMING_ASSETS;
+        public const string STR_MOD_NAME_BASE_GAME = "Base Game";
+
+        public static bool VerboseLog = true;
+
 
         public string BaseGameDir
         {
@@ -31,7 +51,9 @@ namespace ParallelReality
 
         public string ModDir { get; private set; } = string.Empty;
 
-        public bool IsApplyInProgress { get; private set; } = false;
+        public ModManagerState State { get => m_State; }
+
+        //public bool IsApplyInProgress { get; private set; } = false;
         public bool IsGameModded { get => File.Exists(m_ModdedFlagFile); }
 
 
@@ -43,8 +65,44 @@ namespace ParallelReality
         {
             FoundMods = new();
             SelectedMods = new();
+
+            m_FileToModsMap = new();
         }
 
+        public ModManager(string _baseGameDir)
+        {
+            m_BaseGameDir = _baseGameDir;
+            m_ModdedFlagFile = m_BaseGameDir + "/modded.txt";
+            m_ModDir = m_BaseGameDir + "/Mods";
+
+            FoundMods = new();
+            SelectedMods = new();
+
+            m_FileToModsMap = new();
+        }
+
+
+        public ModInfo? GetModWithId(int _id)
+        {
+            foreach (var mod in FoundMods)
+            {
+                if (mod.ModId == _id)
+                    return mod;
+            }
+
+            return null;
+        }
+
+        public ModInfo? GetModWithShortHash(string _shortHash)
+        {
+            foreach (var mod in FoundMods)
+            {
+                if (mod.GetShortHash() == _shortHash)
+                    return mod;
+            }
+
+            return null;
+        }
 
         public List<ModInfo> GetLoadedMods()
         {
@@ -61,6 +119,8 @@ namespace ParallelReality
 
         public void CheckForDownloadedMods()
         {
+            m_State = ModManagerState.CheckingForMods;
+
             Directory.CreateDirectory(ModDir);
 
             FoundMods.Clear();
@@ -69,7 +129,8 @@ namespace ParallelReality
             IEnumerable<string> dirs = Directory.EnumerateDirectories(ModDir);
             foreach (var dir in dirs)
             {
-                if (dir.EndsWith("Base Game"))
+                string? folderName = Path.GetFileName(dir);
+                if (folderName == null || folderName.StartsWith(STR_MOD_NAME_BASE_GAME))
                     continue;
 
                 ModInfo mod = new(dir, index);
@@ -78,54 +139,101 @@ namespace ParallelReality
             }
 
             Console.WriteLine("Found total " + FoundMods.Count + " mods.");
-        }
 
-        public void CheckForAppliedMods()
-        {
-            if (!IsGameModded)
+            string msg = "";
+            string cap = "-----------------------------------------------------------\r\n";
+            if (FoundMods.Count > 0 && VerboseLog)
             {
-                foreach (var mod in FoundMods)
-                {
-                    mod.LoadedOrder = -1;
-                }
-
-                return;
+                msg = cap + "|   Id |   Hash   | Name                           | Load |\r\n" + cap;
             }
 
-            List<ModInfo> loadedMods = new();
-            string[] lines = File.ReadAllLines(m_ModdedFlagFile);
-            foreach (string line in lines)
+            List<ModInfoSimple> loadedModsPrecheck = CheckForAppliedMod_Precheck();
+
+            foreach (var mod in FoundMods)
             {
-                int pos = line.IndexOf('\t');
-                if (pos == -1 || pos + 9 > line.Length)
-                    continue;
-
-                if (!int.TryParse(line[0..pos], out int order))
-                    continue;
-
-                string hash = line[(pos + 1)..(pos + 9)];
-
-                string name = line[(pos + 1)..];
-
-                // ==========
-                foreach (var mod in FoundMods)
+                foreach (var file in mod.ModFiles)
                 {
-                    if (mod.GetShortHash() == hash)
+                    if (!m_FileToModsMap.ContainsKey(file))
                     {
-                        mod.LoadedOrder = order;
-                        loadedMods.Add(mod);
+                        m_FileToModsMap[file] = new List<int>();
+                    }
+
+                    m_FileToModsMap[file].Add(mod.ModId);
+                }
+
+                string hash = mod.GetShortHash();
+                string loadedOrder = "-";
+
+                foreach (var precheck in loadedModsPrecheck)
+                {
+                    if (hash == precheck.Hash)
+                    {
+                        mod.LoadedOrder = precheck.LoadedOrder;
+                        loadedOrder = precheck.LoadedOrder.ToString();
                         break;
                     }
                 }
 
+                msg += string.Format("| {0,4} | {1} | {2} | {3,4} |\r\n",
+                    mod.ModId, mod.GetShortHash(), Utils.ClipString(mod.Name, 30),
+                    loadedOrder);
+                // TODO: at this point LoadedOrder is not populated yet;
             }
 
-            Console.WriteLine("    In which, " + loadedMods.Count + " mods are already loaded.");
-            foreach (var mod in loadedMods)
+            if (FoundMods.Count > 0 && VerboseLog)
             {
-                Console.WriteLine("        '" + mod.Name + "' (Id = " + mod.Index + ")");
+                Console.Write(msg + cap);
+                Console.WriteLine("                                                    ↑ " + loadedModsPrecheck.Count + " mods loaded");
             }
+
+            m_State = ModManagerState.Idle;
         }
+
+        //public void CheckForAppliedMods()
+        //{
+        //    //Console.WriteLine("                                                    ↑ " + 0 + " mods loaded");
+        //    List<ModInfoPrecheck> precheck = CheckForAppliedMod_Precheck();
+        //    if (precheck.Count == 0)
+        //        return;
+
+        //    List<ModInfo> loadedMods = new();
+        //    string[] lines = File.ReadAllLines(m_ModdedFlagFile);
+        //    foreach (string line in lines)
+        //    {
+        //        int pos = line.IndexOf('\t');
+        //        if (pos == -1 || pos + 9 > line.Length)
+        //            continue;
+
+        //        if (!int.TryParse(line[0..pos], out int order))
+        //            continue;
+
+        //        string hash = line[(pos + 1)..(pos + 9)];
+
+        //        string name = line[(pos + 1)..];
+
+        //        // ==========
+        //        foreach (var mod in FoundMods)
+        //        {
+        //            if (mod.GetShortHash() == hash)
+        //            {
+        //                mod.LoadedOrder = order;
+        //                loadedMods.Add(mod);
+        //                break;
+        //            }
+        //        }
+
+        //    }
+
+        //    if (precheck.Count > 0)
+        //    {
+        //        Console.WriteLine("                                                    ↑ " + precheck.Count + " mods loaded");
+        //    }
+        //    //Console.WriteLine("    In which, " + loadedMods.Count + " mods are already loaded.");
+        //    //foreach (var mod in loadedMods)
+        //    //{
+        //    //    Console.WriteLine("        '" + mod.Name + "' (Id = " + mod.ModId + ")");
+        //    //}
+        //}
 
         public void BackupBaseGame(bool _force = false)
         {
@@ -140,20 +248,32 @@ namespace ParallelReality
                 return;
             }
 
-            string part = "/Reality Break_Data/StreamingAssets";
-            string srcDir = BaseGameDir + part;
-            string dstDir = ModDir + "/Base Game" + part;
+            m_State = ModManagerState.BackingUpBaseGame;
+
+            string srcDir = BaseGameDir + "/" + STR_PATH_DATA_STREAMING_ASSETS;
+            string dstDir = ModDir + "/Base Game/" + STR_PATH_DATA_STREAMING_ASSETS;
             CopyFiles(srcDir, dstDir);
 
             Console.WriteLine("Base Game data backed up.");
+
+            m_State = ModManagerState.Idle;
         }
 
-        const string c_Path_StreamingAssets = "/Reality Break_Data/StreamingAssets";
-        public void RestoreBaseGame()
+        public void RestoreBaseGame(Action<string> _callbackUpdateStatus)
         {
-            string srcDir = ModDir + "/Base Game" + c_Path_StreamingAssets;
-            string dstDir = BaseGameDir + c_Path_StreamingAssets;
-            CopyFiles(srcDir, dstDir);
+            string srcDir = ModDir + "/Base Game/" + STR_PATH_DATA_STREAMING_ASSETS;
+            string dstDir = BaseGameDir + "/" + STR_PATH_DATA_STREAMING_ASSETS;
+
+            m_State = ModManagerState.RestoringBaseGame;
+
+            //_callbackUpdateStatus("Deleting Modded files..");
+
+            _callbackUpdateStatus("Restoring Base Game files..");
+#if DEV
+            Thread.Sleep(1500);
+#else
+            CopyFiles(srcDir, dstDir, true);
+#endif
 
             if (File.Exists(m_ModdedFlagFile))
             {
@@ -166,17 +286,21 @@ namespace ParallelReality
             }
 
             Console.WriteLine("Base Game data restored.");
+
+            m_State = ModManagerState.Idle;
         }
 
-        public void ApplyMods(List<int> _modIndices, Action<int, int> _callbackReportProgress, CancellationToken _cancellationToken)
+        public void ApplyMods(List<int> _modIndices, Action<string> _callbackUpdateStatus, CancellationToken _cancellationToken)
         {
-            IsApplyInProgress = true;
+            m_State = ModManagerState.ApplyingMods;
+
+            //IsApplyInProgress = true;
 
             CopiesStatus.TotalFiles = 0;
             for (int i = 0; i < _modIndices.Count; ++i)
             {
                 ModInfo mod = FoundMods[_modIndices[i]];
-                CopiesStatus.TotalFiles += mod.ModifiedFiles.Count;
+                CopiesStatus.TotalFiles += mod.ModFiles.Count;
             }
 
             Console.WriteLine("Applying " + _modIndices.Count + " mods (total " + CopiesStatus.TotalFiles + " files)..");
@@ -188,50 +312,104 @@ namespace ParallelReality
 
                 try
                 {
-                    ApplyMod(mod, _callbackReportProgress, _cancellationToken);
+                    ApplyMod(mod, _callbackUpdateStatus, _cancellationToken);
                 }
                 catch (OperationCanceledException)
                 {
-                    IsApplyInProgress = false;
+                    m_State = ModManagerState.Idle;
+                    //IsApplyInProgress = false;
                     CopiesStatus.TotalFiles = 0;
                     CopiesStatus.CopiedFiles = 0;
 
+                    Console.WriteLine("    User cancelled.");
                     throw;
                 }
 
+                mod.LoadedOrder = i;
+
                 names[i] = i + "\t" + mod.GetShortHash() + "\t" + mod.Name;
-                Console.WriteLine("    Applied mod '" + mod.Name + " (Id = " + mod.Index + ").");
+                Console.WriteLine("    Applied mod '" + mod.Name + " (Id = " + mod.ModId + ").");
             }
 
             File.WriteAllLines(m_ModdedFlagFile, names.ToArray());
 
-            IsApplyInProgress = false;
+            m_State = ModManagerState.Idle;
+            //IsApplyInProgress = false;
             CopiesStatus.TotalFiles = 0;
             CopiesStatus.CopiedFiles = 0;
         }
 
-        private void ApplyMod(ModInfo _mod, Action<int, int> _callbackReportProgress, CancellationToken _cancellationToken)
+        private void ApplyMod(ModInfo _mod, Action<string> _callbackUpdateStatus, CancellationToken _cancellationToken)
         {
-            string modPath = _mod.ModDir + "/" + STR_DATA_FOLDER_NAME + "/";
-            string basePath = BaseGameDir + "/" + STR_DATA_FOLDER_NAME + "/";
-            foreach (string name in _mod.ModifiedFiles)
+            string modPath = _mod.ModDir + "/" + STR_FOLDER_NAME_DATA + "/";
+            string basePath = BaseGameDir + "/" + STR_FOLDER_NAME_DATA + "/";
+            foreach (string name in _mod.ModFiles)
             {
                 _cancellationToken.ThrowIfCancellationRequested();
 
                 string srcName = modPath + name;
                 string dstName = basePath + name;
 
-                File.Copy(srcName, dstName, true);
-                //Thread.Sleep(500);
+#if DEV
+                Thread.Sleep(500);
                 ++CopiesStatus.CopiedFiles;
-                _callbackReportProgress(CopiesStatus.CopiedFiles, CopiesStatus.TotalFiles);
+#else
+                if (File.Exists(srcName))
+                {
+                    File.Copy(srcName, dstName, true);
+                    ++CopiesStatus.CopiedFiles;
+                }
+                else
+                {
+                    Console.WriteLine("File '" + srcName + "' no longer exist.");
+                }
+#endif
+
+                string text = "Applying mods: " + CopiesStatus.CopiedFiles + "/" + CopiesStatus.TotalFiles + " files copied.";
+                _callbackUpdateStatus(text);
+            }
+        }
+
+        private List<ModInfoSimple> CheckForAppliedMod_Precheck()
+        {
+            List<ModInfoSimple> list = new();
+            if (!IsGameModded)
+            {
+                foreach (var mod in FoundMods)
+                {
+                    mod.LoadedOrder = -1;
+                }
+
+                return list;
             }
 
-            //CopyFiles(srcDir, dstDir);
+            string[] lines = File.ReadAllLines(m_ModdedFlagFile);
+            foreach (string line in lines)
+            {
+                int pos = line.IndexOf('\t');
+                if (pos == -1 || pos + 9 > line.Length)
+                    continue;
+
+                if (!int.TryParse(line[0..pos], out int order))
+                    continue;
+
+                string hash = line[(pos + 1)..(pos + 9)];
+
+                string name = line[(pos + 1)..];
+
+                list.Add(new()
+                {
+                    LoadedOrder = order,
+                    Hash = hash,
+                    Name = name,
+                });
+            }
+
+            return list;
         }
 
 
-        private List<string> GetFiles(string _path)
+        private static List<string> GetFiles(string _path)
         {
             List<string> list = new();
 
@@ -248,7 +426,7 @@ namespace ParallelReality
             return list;
         }
 
-        private void CopyFiles(string _srcDir, string _dstDir)
+        private static void CopyFiles(string _srcDir, string _dstDir, bool _overwrite = false)
         {
             _ = Directory.CreateDirectory(_dstDir);
 
@@ -256,23 +434,29 @@ namespace ParallelReality
             foreach (string file in files)
             {
                 string dstFilename = _dstDir + "/" + Path.GetFileName(file);
-                File.Copy(file, dstFilename, true);
+                File.Copy(file, dstFilename, _overwrite);
             }
 
             string[] dirs = Directory.GetDirectories(_srcDir);
             foreach (string dir in dirs)
             {
                 string dstDirName = _dstDir + "/" + Path.GetFileName(dir);
-                CopyFiles(dir, dstDirName);
+                CopyFiles(dir, dstDirName, _overwrite);
             }
         }
 
 
+        
 
-        private string m_BaseGameDir = string.Empty;
+        private  string m_BaseGameDir = string.Empty;
+        private  string m_ModDir = string.Empty;
 
-        private string m_ModdedFlagFile = string.Empty;
+        private  string m_ModdedFlagFile = string.Empty;
 
+        private ModManagerState m_State = ModManagerState.Idle;
+
+
+        private readonly Dictionary<string, List<int>> m_FileToModsMap;
     }
 
 }
